refactor: extract shared docs prerendering
