+++ docs/guides/package.json
@@ -26,6 +26,7 @@
     "remark-parse": "^11.0.0",
     "remark-rehype": "^11.1.2",
     "remix": "workspace:^",
+    "remix-docs-shared": "workspace:*",
     "shiki": "^3.22.0",
     "unified": "^11.0.5",
     "unist-util-visit": "^5.1.0"
