+++ docs/shared/prerender/bench/fixtures/blog-post.html
