+++ docs/shared/prerender/bench/src/html-parser.bench.ts
