+++ docs/shared/prerender/utils.ts
